+++ src/main/java/net/jodah/failsafe/internal/CircuitState.java
@@ -15,10 +15,9 @@
  */
 package net.jodah.failsafe.internal;
 
+import net.jodah.failsafe.CircuitBreaker;
 import net.jodah.failsafe.CircuitBreaker.State;
 import net.jodah.failsafe.ExecutionContext;
-import net.jodah.failsafe.internal.util.CircularBitSet;
-import net.jodah.failsafe.util.Ratio;
 
 import java.time.Duration;
 
@@ -28,43 +27,39 @@ import java.time.Duration;
  * @author Jonathan Halterman
  */
 public abstract class CircuitState {
-  static final Ratio ONE_OF_ONE = new Ratio(1, 1);
+  final CircuitBreaker breaker;
+  volatile CircuitStats stats;
 
-  protected CircularBitSet bitSet;
+  CircuitState(CircuitBreaker breaker, CircuitStats stats) {
+    this.breaker = breaker;
+    this.stats = stats;
+  }
 
   public abstract boolean allowsExecution();
-
-  public abstract State getInternals();
 
   public Duration getRemainingDelay() {
     return Duration.ZERO;
   }
 
-  public int getFailureCount() {
-    return bitSet.negatives();
+  public CircuitStats getStats() {
+    return stats;
   }
 
-  public Ratio getFailureRatio() {
-    return bitSet.negativeRatio();
+  public abstract State getState();
+
+  public synchronized void recordFailure(ExecutionContext context) {
+    stats.recordFailure();
+    checkThreshold(context);
   }
 
-  public int getSuccessCount() {
-    return bitSet.positives();
+  public synchronized void recordSuccess() {
+    stats.recordSuccess();
+    checkThreshold(null);
   }
 
-  public Ratio getSuccessRatio() {
-    return bitSet.positiveRatio();
+  public void handleConfigChange() {
   }
 
-  public void recordFailure(ExecutionContext context) {
-  }
-
-  public void recordSuccess() {
-  }
-
-  public void setFailureThreshold(Ratio threshold) {
-  }
-
-  public void setSuccessThreshold(Ratio threshold) {
+  void checkThreshold(ExecutionContext context) {
   }
 }